server/pyproject.toml
[project]
name = "jsontopydantic"
version = "0.1.0"
description = "[JSON to Pydantic](https://jsontopydantic.com) is a tool that lets you convert JSON objects into Pydantic models."
authors = [{ name = "Michael Wooley", email = "william.wooley@motional.com" }]
dependencies = [
    "fastapi>=0.85.1",
    "pydantic[dotenv,email]>=1.10.2",
    "datamodel-code-generator>=0.13.3",
    "mangum>=0.16.0",
    "genson>=1.2.2",
    "pyhumps>=3.8.0",
]
requires-python = ">=3.10"
license = { text = "MIT" }
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.10.0",
    "uvicorn[standard]>=0.19.0",
    "ipython>=8.5.0",
]
test = ["pytest>=7.2.0"]
codegen = [
    "jinja2>=3.1.2",
]

[tool.pdm.scripts]
dev = "uvicorn app.main:app --reload"
web = { shell = "cd ../client && npm run dev" }
